requirements.txt
blinker==1.7.0
click==8.1.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.2
